Added game over screen the camera now follows the player.
Added a game manager which takes care of events that affect the whole game, like when you fail or success a level by example.
Added a spike that can cause a game over if collided with the player.
Created a script that allows the camera to follow the player (you cannot simply associate the camera with the player because if you do so the camera also follows the rotation of the player which leads to the whole stage rotating from the user's perspective). The camera position also has an offset of -10 on the z axis so that the user actually sees the scene and not from the player's perspective.